backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
alembic==1.16.5
asyncpg==0.30.0
aiosqlite==0.20.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
openai==1.100.2
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.3.0
deepgram-sdk==3.7.7
pyttsx3==2.99
pyaudio==0.2.14
sounddevice==0.5.1
websockets==13.1
